dockerairflow/requirements.txt
pandas==2.2.1
mysql-connector-python==8.3.0
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.3.5